damping and trim
amended the calculation for wing lift calculation to assume 0 angle of yaw as well as no control surface deflection. amended the moment arms to be closer to what is correct for the plane.

updated trim calculation for the new dynamics model and calculated trim. the results are still a bit surprising as at trim (no acceleration, not angular acceleration) there is non-zero fz (lift-grav)